Cambios para autenticación de usuarios; añadidos botones a todas las vistas para poder navegar entre ellas

+++ views/listar.php
@@ -7,7 +7,15 @@
     </head>
     <body>
         <h2>Lista de Videojuegos</h2>
-        <a href="index.php?accion=añadir" class="btn btn-success btn-lg">Añadir nuevo videojuego</a><br>
+
+        <?php if(isset($_SESSION['usuario_id'])): ?>
+            <p>Bienvenido, <?= $_SESSION['usuario_email'] ?></p>
+            <a href="index.php?accion=logout" class="glyphicon glyphicon-off">Cerrar sesión</a><br><br>
+            <a href="index.php?accion=añadir" class="btn btn-success btn-lg">Añadir nuevo videojuego</a><br>
+        <?php else: ?>
+            <a href="index.php?accion=registroUsuario" class="glyphicon glyphicon-user">Registrarse</a><br>
+            <a href="index.php?accion=login" class="glyphicon glyphicon-log-in">Iniciar sesión</a><br>
+        <?php endif; ?>
 
         <div class="container-fluid">
             <table class="table table-striped">
@@ -17,7 +25,9 @@
                         <th>Duración (Horas)</th>
                         <th>Género</th>
                         <th>Info específica</th>
-                        <th>Acciones</th>
+                        <?php if(isset($_SESSION['usuario_id'])): ?>
+                            <th>Acciones</th>
+                        <?php endif; ?>
                     </tr>
                 </thead>
                 <tbody>
@@ -35,10 +45,12 @@
                                 }
                             ?>
                         </td>
-                        <td>
-                            <a href="index.php?accion=editar&id=<?= $juego->getId() ?>" class="btn btn-warning btn-sm">Editar</a>
-                            <a href="index.php?accion=eliminar&id=<?= $juego->getId() ?>" onclick="return confirm('Eliminar este videojuego?')" class="btn btn-danger btn-sm">Eliminar</a>
-                        </td>
+                        <?php if(isset($_SESSION['usuario_id'])): ?>
+                            <td>
+                                <a href="index.php?accion=editar&id=<?= $juego->getId() ?>" class="btn btn-warning btn-sm">Editar</a>
+                                <a href="index.php?accion=eliminar&id=<?= $juego->getId() ?>" onclick="return confirm('Eliminar este videojuego?')" class="btn btn-danger btn-sm">Eliminar</a>
+                            </td>
+                        <?php endif; ?>
                     </tr>
                     <?php endforeach; ?>
                 </tbody>
